feat: add tests for addReleaseNoteTemplate and refactor file structure
