+++ src/CLI.ts
@@ -1,9 +1,9 @@
 #!/usr/bin/env node
 
 import { Command, Option } from "commander";
-import { addReleaseNoteTemplate } from "./addReleaseNoteTemplate.js";
 import { checkNpmTestCompletion } from "./checkNpmTestCompletion.js";
 import { cleanMerged } from "./cleanMerged.js";
+import { addReleaseNoteTemplate } from "./init/addReleaseNoteTemplate.js";
 import { initLabel } from "./initLabel.js";
 import { postversion } from "./postversion.js";
 import { preversion } from "./preversion.js";
